requirements.txt
aiohttp==3.5.4
alembic==1.0.11
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
more-itertools==7.2.0
multidict==4.5.2
packaging==19.2
pluggy==0.13.1
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.0
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.5
tzlocal==1.5.1
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.15.4
yarl==1.3.0
zipp==0.6.0
